serverapp/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
azure-cosmos==4.5.1
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
dj-database-url==2.1.0
